+++ frontend/src/components/analystTools/GCDocumentsComparisonTool.js
@@ -1,93 +1,72 @@
 import React, {useCallback, useEffect, useState} from 'react';
 import propTypes from 'prop-types';
+import { Collapse } from 'react-collapse';
 import styled from 'styled-components';
 import { makeStyles } from '@material-ui/core/styles';
 import TextField from '@material-ui/core/TextField';
-import {Dialog, DialogActions, DialogContent, DialogTitle, Grid, Typography, Collapse} from '@material-ui/core';
-import { TransitionGroup } from 'react-transition-group';
+import CheckBoxOutlineBlankIcon from '@material-ui/icons/CheckBoxOutlineBlank';
+import { 
+	Grid, 
+	Typography, 
+	Checkbox,
+	FormControl,
+	InputLabel,
+	MenuItem,  
+	Select
+} from '@material-ui/core';
+import {trackEvent} from '../telemetry/Matomo';
 import GCAnalystToolsSideBar from './GCAnalystToolsSideBar';
 import GameChangerAPI from '../api/gameChanger-service-api';
-import {setState} from '../../utils/sharedFunctions';
+import { setState, handleSaveFavoriteDocument } from '../../utils/sharedFunctions';
 import LoadingIndicator from '@dod-advana/advana-platform-ui/dist/loading/LoadingIndicator';
 import {gcOrange} from '../common/gc-colors';
 import {
 	encode, 
 	handlePdfOnLoad,
 	getOrgToOrgQuery,
-	getTypeQuery
+	getTypeQuery,
+	getTrackingNameForFactory,
+	exportToCsv,
+	convertDCTScoreToText
 } from '../../utils/gamechangerUtils';
-import {Card} from '../cards/GCCard';
-import CloseIcon from '@material-ui/icons/Close';
 import GCTooltip from '../common/GCToolTip';
 import GCButton from '../common/GCButton';
+import ExportIcon from '../../images/icon/Export.svg';
 
 const _ = require('lodash');
 
 const gameChangerAPI = new GameChangerAPI();
 
-const PAGE_SIZE = 20
-
-const CloseButton = styled.div`
-	padding: 6px;
-	background-color: white;
-	border-radius: 5px;
-	color: #8091a5 !important;
-	border: 1px solid #b0b9be;
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	flex: 0.4;
-	position: absolute;
-	right: 15px;
-	top: 15px;
-`;
-
-const DocumentCompareContainer = styled.div`
-	height: 500px;
-
-	.relevant-doc-text-container {
-		display: flex;
-		place-content: space-evenly;
-		width: 100%;
-		
-		& .relevant-doc-text {
-			color: ${'#000000DE'}
-			font-family: Montserrat;
-			font-size: 14px;
-			margin-bottom: 20px;
-			width: 48%;
-		}
+const styles = {
+	image: {
+		display: 'flex',
+		justifyContent: 'center',
+		margin: 'auto',
+		marginLeft: '10px',
+		height: 30,
+		color: '#939395',
+		cursor: 'pointer'
+	},
+	checkbox: {
+		padding: '9px'
+	},
+	collapsedInput: {
+		margin: 'auto 0',
+		display: '-webkit-box',
+		WebkitLineClamp: 3,
+		WebkitBoxOrient: 'vertical',
+		overflow: 'hidden',
+	},
+	resultsText:{
+		fontSize: 24,
+		fontFamily: 'Montserrat',
+		borderBottom: '2px solid #BCCBDB',
+		display: 'flex',
+		placeContent: 'space-between',
+		marginTop: 20,
+		marginLeft: 20
 	}
-	
-	.compare-area {
-		display: flex;
-		place-content: space-evenly;
-		margin: auto;
-		width: 100%;
-		height: 740px;
-	
-		& .uploaded-doc {
-			background-color: ${'#ffffff'};
-	        border: 2px solid ${'#BCCBDB'};
-	        border-radius: 6px;
-	        width: 48%;
-	        padding: 10px;
-	        
-	        & .compare-header {
-		        color: ${'#000000DE'};
-		        font-size: 16px;
-		        font-family: Montserrat;
-		        font-weight: bold;
-		        margin-bottom: 10px;
-	        }
-		}
-		
-		& .relevant-document {
-			width: 48%;
-		}
-	}
-`;
+}
 
 const DocumentInputContainer = styled.div`
 	border: 5px ${'#FFFFFF'};
@@ -163,38 +142,6 @@ const DocumentInputContainer = styled.div`
 	
 `;
 
-const SimilarDocumentsContainer = styled.div`
-	margin: 20px 0px 0px 20px;
-
-	.displaying-results-text {
-		font-size: 24px;
-		font-family: Montserrat;
-		border-bottom: 2px solid ${'#BCCBDB'};
-		display: flex;
-		place-content: space-between;
-		margin-top: 5px;
-		
-		& .text {
-			margin: auto 0px;
-			
-			& span {
-				font-weight: bold;
-			}
-		}
-		
-		& .gcPagination {
-			& .pagination {
-				margin: unset;
-				font-size: 12px;
-			}
-		}
-	}
-	
-	.results-container {
-		margin-top: 5px;
-	}
-`;
-
 const getPresearchData = async (state, dispatch) => {
 	const { cloneData } = state;
 	if (_.isEmpty(state.presearchSources)) {
@@ -237,6 +184,7 @@ const getPresearchData = async (state, dispatch) => {
 const resetAdvancedSettings = (dispatch) => {
 	dispatch({type: 'RESET_ANALYST_TOOLS_SEARCH_SETTINGS'});
 }
+
 const GCDocumentsComparisonTool = (props) => {
 	
 	const classes = useStyles();
@@ -255,17 +203,45 @@ const GCDocumentsComparisonTool = (props) => {
 	
 	const [paragraphText, setParagraphText] = useState('');
 	const [paragraphs, setParagraphs] = useState([]);
+	const [selectedInput, setSelectedInput] = useState(undefined);
 	const [returnedDocs, setReturnedDocs] = useState([]);
 	const [viewableDocs, setViewableDocs] = useState([]);
 	const [loading, setLoading] = useState(false);
-	const [page, setPage] = useState(1);
 	const [compareDocument, setCompareDocument] = useState(undefined);
-	const [compareParagraphIndex, setCompareParagraphIndex] = useState(0);
+	const [selectedParagraph, setSelectedParagraph] = useState(undefined);
+	const [itemsToCombine, setItemsToCombine] = useState({});
+	const [combineDisabled, setCombineDisabled] = useState(true);
 	const [filtersLoaded, setFiltersLoaded] = useState(false);
 	const [noResults, setNoResults] = useState(false);
 	const [filterChange, setFilterChange] = useState(false);
-	const [highlightList, setHighlightList] = useState([]);
-	const [highlightIndex, setHighlightindex] = useState(0);
+	const [inputError, setInputError] = useState(false);
+	const [collapseKeys, setCollapseKeys] = useState([]);
+	const [feedbackList, setFeedbackList] = useState({});
+	const [sortType, setSortType] = useState('Similarity Score');
+	const [needsSort, setNeedsSort] = useState(true);
+	const [sortOrder, setSortOrder] = useState('desc');
+
+	const handleSetParagraphs = useCallback(() => {
+		const paragraphs = paragraphText.split('\n').map((paragraph, idx) => {
+			return {text: paragraph.trim(), id: idx};
+		}).filter(paragraph => paragraph.text.length > 0);
+
+		if(paragraphs.length > 5){
+			setInputError(true);
+		}else{
+			setInputError(false);
+		}
+		
+		setParagraphs(paragraphs);
+	}, [paragraphText])
+
+	useEffect(() => {
+		let disable = 0;
+		Object.keys(itemsToCombine).forEach(par => {
+			if(itemsToCombine[par]) disable++;
+		})
+		setCombineDisabled(disable < 2)
+	}, [itemsToCombine])
 	
 	useEffect(() => {
 		if(!filtersLoaded){
@@ -285,12 +261,7 @@ const GCDocumentsComparisonTool = (props) => {
 	useEffect(() => {
 		if (state.runDocumentComparisonSearch) {
 			setLoading(true);
-			
-			const paragraphs = paragraphText.split('\n').map(paragraph => {
-				return paragraph.trim();
-			}).filter(paragraph => paragraph.length > 0);
-			
-			setParagraphs(paragraphs);
+			setCollapseKeys([]);
 
 			const filters = {
 				orgFilters: getOrgToOrgQuery(allOrgsSelected, orgFilter),
@@ -300,7 +271,22 @@ const GCDocumentsComparisonTool = (props) => {
 			}
 			
 			gameChangerAPI.compareDocumentPOST({cloneName: state.cloneData.cloneName, paragraphs: paragraphs, filters}).then(resp => {
-				if(resp.data.docs.length <= 0) setNoResults(true);
+				if(resp.data.docs.length <= 0) {
+					setNoResults(true)
+				}else{
+					let paragraph;
+					const document = resp.data.docs.find(doc => {
+						const foundPar = doc.paragraphs.find(par => par.paragraphIdBeingMatched === selectedInput)
+						if(foundPar){
+							paragraph = foundPar;
+							return true;
+						}else {
+							return false;
+						}
+					})
+					setCompareDocument(document);
+					setSelectedParagraph(paragraph);
+				}
 				setReturnedDocs(resp.data.docs);
 				setState(dispatch, {runDocumentComparisonSearch: false});
 				setLoading(false);
@@ -312,73 +298,85 @@ const GCDocumentsComparisonTool = (props) => {
 			});
 		}
 		
-	}, [state.runDocumentComparisonSearch, paragraphText, state.cloneData.cloneName, dispatch, allOrgsSelected, orgFilter, allTypesSelected, typeFilter, publicationDateFilter, includeRevoked]);
+	}, [state.runDocumentComparisonSearch, paragraphText, state.cloneData.cloneName, dispatch, allOrgsSelected, orgFilter, allTypesSelected, typeFilter, publicationDateFilter, includeRevoked, paragraphs, selectedInput]);
 	
 	useEffect(() => {
-		setViewableDocs(returnedDocs)
-	}, [returnedDocs]);
-	
-	useEffect(() => {
-		if (state.compareModalOpen) {
-			const doc = returnedDocs.filter(document => {
-				return document.filename === state.compareFilename;
-			})[0];
-			
-			let tmpCompareIdx;
-			doc.paragraphs.forEach(par => {
-				if (tmpCompareIdx === undefined) {
-					tmpCompareIdx = par.paragraphIdBeingMatched;
-				}
-			});
-			
-			setCompareParagraphIndex(tmpCompareIdx);
-			const highlights = doc.paragraphs.filter(p => p.paragraphIdBeingMatched === tmpCompareIdx);
-			setHighlightList(highlights);
-			setCompareDocument(doc);
+		if(needsSort && returnedDocs.length){
+			setNeedsSort(false);
+			const newViewableDocs = returnedDocs.filter(doc => {
+				return doc.paragraphs.find(match => match.paragraphIdBeingMatched === selectedInput);
+			})
+			const order = sortOrder === 'desc' ? 1 : -1;
+			let sortFunc = () => {};
+			switch(sortType){
+				case 'Alphabetically':
+					sortFunc = (docA, docB) => {
+						if(docA.title > docB.title) return order;
+						if(docA.title < docB.title) return -order;
+						return 0;
+					}
+					break;
+				case 'Date Published':
+					sortFunc = (docA, docB) => {
+						if(docA.publication_date_dt > docB.publication_date_dt) return -order;
+						if(docA.publication_date_dt < docB.publication_date_dt) return order;
+						if(!docA.publication_date_dt) return 1;
+						if(!docB.publication_date_dt) return -1;
+						return 0;
+					}
+					break;
+				default: 
+					sortFunc = (docA, docB) => {
+						if(docA.score > docB.score) return -order;
+						if(docA.score < docB.score) return order;
+						return 0;
+					}
+					break;
+			}
+			const sortedDocs = newViewableDocs.sort(sortFunc);
+			if(!selectedParagraph) {
+				setSelectedParagraph(sortedDocs[0].paragraphs[0]);
+				setCompareDocument(sortedDocs[0]);
+			}
+			setViewableDocs(sortedDocs);
 		}
-	}, [returnedDocs, state.compareModalOpen, state.compareFilename]);
+	}, [needsSort, returnedDocs, selectedInput, sortType, viewableDocs, sortOrder, selectedParagraph]);
 
-	useEffect(() => {
-		if(state.ignoredDocs.length){
-			const { item, index } = state.ignoredDocs[0];
-			const searchedParagraph = paragraphs[item.paragraphs[index].paragraphIdBeingMatched]
-			const matchedParagraphId = item.paragraphs[index].id;
+	const handleFeedback = (doc, paragraph, positiveFeedback) => {
+		let undo = false;
+		if(positiveFeedback === feedbackList[paragraph.id]) {
+			undo = true;
+			const newList = {...feedbackList};
+			delete newList[paragraph.id];
+			setFeedbackList(newList);
+		} else {
+			setFeedbackList({...feedbackList, [paragraph.id]: positiveFeedback});
+		};
 
-			gameChangerAPI.compareFeedbackPOST({
-				searchedParagraph,
-				matchedParagraphId,
-				docId: item.id,
-				positiveFeedback: false
-			});
-			setState( dispatch, {ignoredDocs: []})
+		const searchedParagraph = paragraphs.find(input => input.id === paragraph.paragraphIdBeingMatched).text;
+		const matchedParagraphId = paragraph.id;
 
-			const newViewableDocs = viewableDocs;
-			const docIndex = viewableDocs.findIndex((doc) => item.filename === doc.filename);
-			newViewableDocs[docIndex].paragraphs.splice(index, 1);
-			if(!newViewableDocs[docIndex].paragraphs.length) newViewableDocs.splice(docIndex, 1);
-			setViewableDocs(newViewableDocs);
-		}
-	}, [state.ignoredDocs, viewableDocs, dispatch, paragraphs])
+		gameChangerAPI.compareFeedbackPOST({
+			searchedParagraph,
+			matchedParagraphId,
+			docId: doc.id,
+			positiveFeedback,
+			undo
+		})
+	}
 	
 	const measuredRef = useCallback(
 		(node) => {
 			if (node !== null && compareDocument) {
-				
-				const matchingPars = compareDocument.paragraphs.filter(par => {
-					return par.paragraphIdBeingMatched === compareParagraphIndex;
-				});
 
-				const isDLA = compareDocument.display_org_s === 'Defense Logistics Agency';
-
-				if (compareDocument && matchingPars.length > 0) {
+				if (compareDocument && selectedParagraph) {
 					gameChangerAPI
 						.dataStorageDownloadGET(
 							encode(compareDocument.filename || ''),
-							`"${highlightList[highlightIndex].par_raw_text_t}"`,
-							matchingPars[0].page_num_i + 1,
+							`"${selectedParagraph.par_raw_text_t}"`,
+							selectedParagraph.page_num_i + 1,
 							true,
-							state.cloneData,
-							isDLA
+							state.cloneData
 						)
 						.then((url) => {
 							node.src = url;
@@ -386,155 +384,234 @@ const GCDocumentsComparisonTool = (props) => {
 				}
 			}
 		},
-		[compareDocument, state.cloneData, compareParagraphIndex, highlightList, highlightIndex]
+		[compareDocument, state.cloneData, selectedParagraph]
 	);
-	
-	const getMatchingParsCount = (compareIdx) => {
-		return compareDocument.paragraphs.filter(par => {
-			return par.paragraphIdBeingMatched === compareIdx;
-		}).length;
-	}
-	
-	const handleSetCompareIndex = (idx) => {
-		const parCount = getMatchingParsCount(idx);
-		if (parCount > 0){
-			setCompareParagraphIndex(idx);
-			const highlights = compareDocument.paragraphs.filter(p => p.paragraphIdBeingMatched === idx);
-			setHighlightList(highlights);
-			setHighlightindex(highlightIndex >= parCount -1 ? 0 : highlightIndex + 1);
-		}
-	}
+
+	useEffect(() => {
+		handleSetParagraphs();
+	}, [paragraphText, handleSetParagraphs])
 	
 	const reset = () => {
 		setParagraphText('');
+		setInputError(false);
 		setReturnedDocs([]);
 		setViewableDocs([]);
-		setPage(1);
 	}
-	
-	const buildCards = (docs) => {
-		return <TransitionGroup>
-			{_.map(docs, (item, idx) => {
-				item.type = 'document';
-				item.isCompare = true;
-				item.dataToQuickCompareTo = paragraphs;
-				item.pageHits = [];
-				return (
-					<Collapse key={item.filename} timeout={1000}>
-						<Card key={idx}
-							item={item}
-							idx={idx}
-							state={{
-								...state,
-								selectedDocuments: new Map(), componentStepNumbers: {}, listView: true, showSideFilters: false
-							}}
-							dispatch={dispatch}
-						/>
-					</Collapse>
-				);
-			})}
-		</TransitionGroup>
+
+	const handleCheck = (id) => {
+		setItemsToCombine({
+			...itemsToCombine,
+			[id]: !itemsToCombine?.[id],
+		});
+	}
+
+	const removeParagraph = (id) => {
+		const newParagraphs = paragraphs.filter((par) => par.id !== id);
+		if(id === selectedInput) {
+			setSelectedInput(newParagraphs[0].id);
+			setNeedsSort(true);
+			setToFirstResultofInput(newParagraphs[0].id);
+		}
+		setParagraphs(newParagraphs);
+	}
+
+	const handleCombine = () => {
+		const oldParagraphs = [];
+		const paragraphsToCombine = [];
+		paragraphs.forEach(par => {
+			if(itemsToCombine[par.id]){ 
+				paragraphsToCombine.push(par.text)
+			}else{
+				oldParagraphs.push(par.text)
+			}
+		});
+		const combinedParagraphs = paragraphsToCombine.join(' ');
+		oldParagraphs.push(combinedParagraphs);
+		const newParagraphs = oldParagraphs.map((text, idx) => {
+			return {text, id: idx};
+		})
+		setParagraphs(newParagraphs);
+		const newParagraphText = newParagraphs.map(paragraph => paragraph.text).join('\n');
+		setParagraphText(newParagraphText);
+		setReturnedDocs([]);
+		handleCompare();
+	}
+
+	const handleCompare = () => {
+		setNoResults(false);
+		setFilterChange(false);
+		setSelectedInput(paragraphs?.[0].id);
+		setNeedsSort(true);
+		setItemsToCombine({});
+		setState(dispatch, { runDocumentComparisonSearch: true });
+	}
+
+	const saveDocToFavorites = (filename, paragraph) => {
+		const text = paragraphs.find(input => input.id === paragraph.paragraphIdBeingMatched).text;
+		handleSaveFavoriteDocument({filename, is_favorite: true, search_text: text, favorite_summary: `Document Comparison result of "${text}"`}, state, dispatch)
+	}
+
+	const exportSingleDoc = (document) => {
+		const exportList = [];
+		document.paragraphs.forEach(paragraph => {
+			const textInput = paragraphs.find(input => paragraph.paragraphIdBeingMatched === input.id).text
+			exportList.push({
+				filename: document.filename,
+				title: document.title,
+				page: paragraph.page_num_i + 1,
+				textInput,
+				textMatch: paragraph.par_raw_text_t,
+				score: convertDCTScoreToText(paragraph.score)
+			})
+		})
+		heandleExport(exportList, 'ExportSindleDocCSV');
+	};
+
+	const exportAll = () => {
+		const exportList = [];
+		returnedDocs.forEach(document => {
+			document.paragraphs.forEach(paragraph => {
+				const textInput = paragraphs.find(input => paragraph.paragraphIdBeingMatched === input.id).text
+				exportList.push({
+					filename: document.filename,
+					title: document.title,
+					page: paragraph.page_num_i + 1,
+					textInput,
+					textMatch: paragraph.par_raw_text_t,
+					score: convertDCTScoreToText(paragraph.score)
+				})
+			})
+		})
+		heandleExport(exportList, 'ExportSindleDocCSV');
+	};
+
+	const heandleExport = (exportList, type) => {
+		try{
+			trackEvent(
+				getTrackingNameForFactory(state.cloneData.clone_name), 
+				'DocumentComparisonTool', 
+				type, 
+				exportList.length
+			);
+			exportToCsv(
+				'DocumentComparisonData.csv', 
+				exportList, 
+				true
+			);
+		} catch (e) {
+			console.error(e);
+			return [];
+		}
+	}
+
+	const setToFirstResultofInput = (inputId) => {
+		let paragraph;
+		const document = viewableDocs.find(doc => {
+			const foundPar = doc.paragraphs.find(par => par.paragraphIdBeingMatched === inputId)
+			if(foundPar){
+				paragraph = foundPar;
+				return true;
+			}else {
+				return false;
+			}
+		})
+		setCompareDocument(document);
+		setSelectedParagraph(paragraph);
+	}
+
+	const handleSelectInput = (id) => {
+		setSelectedParagraph(undefined);
+		setSelectedInput(id);
+		setNeedsSort(true);
+	}
+
+	const handleChangeOrder = (order) => {
+		setSortOrder(order);
+		setNeedsSort(true);
 	}
 	
 	return (
 		<>
-			<Dialog
-				open={state.compareModalOpen}
-				scroll={'paper'}
-				maxWidth="xl"
-				disableEscapeKeyDown
-				disableBackdropClick
-				classes={{
-					paperWidthXl: classes.dialogXl,
-				}}
-			>
-				<DialogTitle>
-					<CloseButton onClick={() => setState(dispatch, {compareModalOpen: false})}>
-						<CloseIcon fontSize="large" />
-					</CloseButton>
-				</DialogTitle>
-
-				<DialogContent>
-					<DocumentCompareContainer>
-						<div className={'relevant-doc-text-container'}>
-							<Typography className={'relevant-doc-text'}>Click each paragraph below to jump to the page the match was found on</Typography>
-							<Typography className={'relevant-doc-text'}>This document may contain relevant information to your search</Typography>
-						</div>
-						{compareDocument &&
-							<div className={'compare-area'}>
-								<div className={'uploaded-doc'}>
-									<Typography className={'compare-header'}>Uploaded Document</Typography>
-									{paragraphs.map((paragraph, idx) => (
-										<>
-											<div
-												style={{
-													border: idx === compareParagraphIndex ? `2px solid ${'#386f94'}` : '',
-													padding: idx === compareParagraphIndex ? `5px` : '',
-													borderRadius: 6,
-													cursor: getMatchingParsCount(idx) > 0 ? 'pointer' : ''
-												}}
-												onClick={() => handleSetCompareIndex(idx)}
-											>
-												{paragraph}
-											</div>
-											<br/>
-										</>
-									))}
-								</div>
-								<div className={'relevant-document'}>
-									<iframe
-										title={'PDFViewer'}
-										className="aref"
-										id={'pdfViewer'}
-										ref={measuredRef}
-										onLoad={() =>
-											handlePdfOnLoad(
-												'pdfViewer',
-												'viewerContainer',
-												compareDocument.filename,
-												'PDF Viewer'
-											)
-										}
-										style={{width: '100%', height: '100%'}}
-									/>
-								</div>
-							</div>
-						}
-					</DocumentCompareContainer>
-				</DialogContent>
-
-				<DialogActions>
-					<div
-						style={{
-							display: 'flex',
-							justifyContent: 'space-between',
-							width: '100%',
-							margin: '0px 18px',
-						}}
-					>
-					
-					</div>
-				</DialogActions>
-			</Dialog>
-			
 			<Grid container style={{marginTop: 20}}>
 				<Grid item xs={12}>
-					<div style={{fontWeight:'bold', alignItems: 'center', fontFamily: 'Noto Sans',}}>
+					<div  style={{display: 'flex'}}>
+						<div style={{fontWeight:'bold', alignItems: 'center', fontFamily: 'Noto Sans',}}>
 					The Document Comparison Tool enables you to input text and locate policies in the GAMECHANGER policy repository with semantically similar language. Using the Document Comparison Tool below, you can conduct deeper policy analysis and understand how one piece of policy compares to the GAMECHANGER policy repository.
+						</div>
+						{(!loading && returnedDocs.length > 0) &&
+							<div style={{display: 'flex', marginLeft: 20}}>
+								<FormControl variant="outlined" classes={{root:classes.root}} style={{marginLeft: 'auto', margin: '-10px 0px 0px 0px', minWidth: 195}}>
+									<InputLabel classes={{root: classes.formlabel}} id="view-name-select">Sort</InputLabel>
+									<Select
+										className={`MuiInputBase-root`}
+										labelId="re-view-name"
+										label="View"
+										id="re-view-name-select"
+										value={sortType}
+										onChange={(event) => {
+											setSortType(event.target.value);
+											setNeedsSort(true);
+										}}
+										classes={{ root: classes.selectRoot, icon: classes.selectIcon }}
+										autoWidthd
+									>
+										<MenuItem key={`Similarity Score`} value={'Similarity Score'}>Similarity Score</MenuItem>
+										<MenuItem key={`Alphabetically`} value={'Alphabetically'}>Alphabetically</MenuItem>
+										<MenuItem key={`Date Published`} value={'Date Published'}>Date Published</MenuItem>
+									</Select>
+								</FormControl>
+								<div style={{width: '40px', marginLeft: '10px', display: 'flex'}}>
+									<i 
+										className="fa fa-sort-amount-desc"
+										style={{marginTop: '50%', marginRight: '5px', cursor: 'pointer', color: sortOrder === 'desc' ? 'rgb(233, 105, 29)' : 'grey'}} 
+										aria-hidden="true"
+										onClick={() => {handleChangeOrder('desc')}	}
+									></i>
+									<i 
+										className="fa fa-sort-amount-asc"
+										style={{marginTop: '50%', cursor: 'pointer', color: sortOrder === 'asc' ? 'rgb(233, 105, 29)' : 'grey'}} 
+										aria-hidden="true"
+										onClick={() => {handleChangeOrder('asc')}	}
+									></i>
+								</div>
+								<GCTooltip title="Export all results" placement="bottom" arrow>
+									<GCButton
+										onClick={exportAll}
+										style={{
+											minWidth: 50,
+											padding: '0px 7px',
+											height: 50,
+											marginTop: 6
+										}}
+									>
+										<img
+											src={ExportIcon}
+											style={{
+												margin: '0 0 3px 3px',
+												width: 15,
+											}}
+											alt="export"
+										/>
+									</GCButton>
+								</GCTooltip>
+							</div>
+						}
 					</div>
 				</Grid>
-				<Grid item xs={3} style={{marginTop: 20}}>
+				<Grid item xs={2} style={{marginTop: 20}}>
 					<GCAnalystToolsSideBar context={context} />
 					<GCButton 
 						isSecondaryBtn 
 						onClick={() => resetAdvancedSettings(dispatch)}
 						style={{margin: 0, width: '100%'}}
 					>
-						Clear filters
+						Clear Filters
 					</GCButton>
 					{!loading && returnedDocs.length > 0 && <GCButton 
 						onClick={() => { 
 							setNoResults(false);
+							setReturnedDocs([]);
+							setNeedsSort(true);
 							setState(dispatch, { runDocumentComparisonSearch: true });
 						}}
 						style={{margin: '10px 0 0 0', width: '100%'}}
@@ -543,7 +620,8 @@ const GCDocumentsComparisonTool = (props) => {
 						Apply filters
 					</GCButton>}
 				</Grid>
-				<Grid item xs={9}>
+				{!(returnedDocs.length > 0) && !loading &&
+				<Grid item xs={10}>
 					<DocumentInputContainer>
 						<Grid container className={'input-container-grid'} style={{margin: 0}}>
 							<Grid item xs={12}>
@@ -563,7 +641,6 @@ const GCDocumentsComparisonTool = (props) => {
 													multiline
 													rows={1000}
 													variant="outlined"
-													className={classes.inputBoxRoot}
 													value={paragraphText}
 													onChange={(event) => {
 														setParagraphText(event.target.value);
@@ -578,6 +655,18 @@ const GCDocumentsComparisonTool = (props) => {
 													}}
 													placeholder={'Text Content Here'}
 													fullWidth={true}
+													helperText={
+														inputError
+															? 'Input currently limited to five paragraphs'
+															: ''
+													}
+													FormHelperTextProps={{
+														style: {
+															color: 'red',
+															fontSize: 12,
+															backgroundColor: 'rgba(0,0,0,0)'
+														}
+													}}
 												/>
 											</div>
 										</Grid>
@@ -587,55 +676,279 @@ const GCDocumentsComparisonTool = (props) => {
 						</Grid>
 						<Grid container style={{justifyContent:'flex-end'}}>
 							<GCTooltip 
-								title={returnedDocs.length > 0 ? 'Click to start over' : 'Compare Documents'} 
+								title={'Compare Documents'} 
 								placement="top" arrow
 							>
 								<GCButton
 									style={{ marginTop: 20 }}
-									onClick={() => {
-										setNoResults(false);
-										setFilterChange(false);
-										if(!loading && returnedDocs.length > 0) return reset();
-										setState(dispatch, { runDocumentComparisonSearch: true });
-									}}
+									disabled={inputError}
+									onClick={handleCompare}
 								>
-									{!loading && returnedDocs.length > 0 ? 'Reset' : 'Submit'}
+									Submit
 								</GCButton>
 							</GCTooltip>
 						</Grid>
 					</DocumentInputContainer>
-					{loading &&
-					<div style={{display:'flex', justifyContent:'center', flexDirection:'column'}}>
-						<LoadingIndicator customColor={gcOrange}/>
-					</div>
-					}
 					{noResults && !loading && 
-						<SimilarDocumentsContainer>
-							<div className={'displaying-results-text'}>
-								<div className={'text'}>
-									No results found
-								</div>
-							</div>
-						</SimilarDocumentsContainer>
+                        <div style={styles.resultsText}>
+                        	<div className={'text'}>
+                                No results found
+                        	</div>
+                        </div>
 					}
-					{(!loading && returnedDocs.length > 0) &&
-					<>
-						<SimilarDocumentsContainer>
-							<div className={'displaying-results-text'}>
-								<div className={'text'}>
-									Showing results {(page - 1) * PAGE_SIZE + 1} - {returnedDocs.length} of: <span>Possibly Relevant Documents</span>
-								</div>
-							</div>
-							
-							<div className={'results-container'}>
-								
-								{buildCards(viewableDocs)}
-							</div>
-							
-						</SimilarDocumentsContainer>
-					</>
-					}
+					
 				</Grid>
+				}
+				{loading &&
+					<Grid item xs={10}>
+						<div style={{display:'flex', justifyContent:'center', flexDirection:'column'}}>
+							<LoadingIndicator customColor={gcOrange}/>
+						</div>
+					</Grid>
+				}
+				{(!loading && returnedDocs.length > 0) &&
+				<>
+					<Grid item xs={6} style={{marginTop: 20}}>
+						<div style={{margin: '0px 20px', height: '800px'}}>
+							<iframe
+								title={'PDFViewer'}
+								className="aref"
+								id={'pdfViewer'}
+								ref={measuredRef}
+								onLoad={() =>
+									handlePdfOnLoad(
+										'pdfViewer',
+										'viewerContainer',
+										compareDocument.filename,
+										'PDF Viewer'
+									)
+								}
+								style={{width: '100%', height: '100%'}}
+							/>
+						</div>
+					</Grid>
+					<Grid item xs={4} style={{marginTop: 20, height: '800px', overflowY: 'scroll', maxWidth: 'calc(33.333333% + 20px)', flexBasis: 'calc((33.333333% + 20px)', paddingLeft: '20px', marginLeft: '-20px'}}>
+						<div 
+							style={{
+								padding: 20,
+								background: '#F6F8FA 0% 0% no-repeat padding-box',
+								border: '1px dashed #707070',
+								display: 'flex',
+								flexDirection: 'column'
+							}}
+						>
+							<Typography variant="body1" style={{marginBottom: 10}}>
+								Paragraph Input
+							</Typography>
+							{paragraphs.map((paragraph) => (
+								<div
+									key={paragraph.id}
+									style={{
+										border: paragraph.id === selectedInput ? 'none' :`2px solid #B6C6D8`,
+										boxShadow: paragraph.id === selectedInput ? '0px 3px 6px #00000029' : 'none',
+										padding: 10,
+										borderRadius: 6,
+										display: 'flex',
+										lineHeight: '20px',
+										marginBottom: 10,
+										cursor: 'pointer',
+										backgroundColor: paragraph.id === selectedInput ? '#DFE6EE' : '#FFFFFF'
+									}}
+									onClick={() => {
+										handleSelectInput(paragraph.id);
+									}}
+								>
+									{paragraphs.length > 1 &&
+										<div 
+											onClick={(event) => {
+												event.stopPropagation();
+												handleCheck(paragraph.id);
+											}}
+											style={{margin: 'auto 0px'}}
+										>
+											<Checkbox
+												checked={itemsToCombine[paragraph.id] ? true : false}
+												classes={{ root: classes.filterBox }}
+												icon={
+													<CheckBoxOutlineBlankIcon
+														style={{ visibility: 'hidden' }}
+													/>
+												}
+												checkedIcon={
+													<i
+														style={{ color: '#E9691D' }}
+														className="fa fa-check"
+													/>
+												}
+											/>
+										</div>
+									}
+									<div style={paragraph.id === selectedInput ? {margin: 'auto 0'} : styles.collapsedInput}>
+										{paragraph.text}
+									</div>
+									{paragraphs.length > 1 &&
+										<div style={{margin: 'auto 0 auto auto'}}>
+											<i 
+												style={styles.image}
+												onClick={(event) => {
+													event.stopPropagation();
+													removeParagraph(paragraph.id);
+												}} 
+												className="fa fa-trash fa-2x" 
+											/>
+										</div>
+									}
+								</div>
+							))}
+							<div style={{display: 'flex', justifyContent: 'flex-end'}}>
+								<GCButton
+									style={{ marginTop: 0, width: 'fit-content'}}
+									isSecondaryBtn
+									disabled={combineDisabled}
+									onClick={() => {
+										handleCombine();
+									}}
+								>
+									Combine
+								</GCButton>
+								<GCButton
+									style={{ marginTop: 0, width: 'fit-content'}}
+									isSecondaryBtn
+									onClick={() => {
+										setNoResults(false);
+										setFilterChange(false);
+										return reset();
+									}}
+								>
+									Reset
+								</GCButton>
+							</div>
+						</div>
+						<div style={{ marginTop: 20 }}>
+							{
+								viewableDocs.length ? 
+									<></> 
+									: 
+									<div style={{...styles.resultsText, marginLeft: 0}}>
+										<div className={'text'}>
+											No results for paragraph found
+										</div>
+									</div>
+							}
+							{viewableDocs.map((doc) => {
+								const docOpen = collapseKeys[doc.filename] ? collapseKeys[doc.filename] : false;
+								const displayTitle = doc.title;
+								return (
+									<div key={doc.id}>
+										<div
+											className="searchdemo-modal-result-header"
+											style={{ marginTop: 0 }}
+											onClick={(e) => {
+												e.preventDefault();
+												setCollapseKeys({ ...collapseKeys, [doc.filename]: !docOpen });
+											}}
+										>
+											<i
+												style={{
+													marginRight: docOpen ? 10 : 14,
+													fontSize: 20,
+													cursor: 'pointer',
+												}}
+												className={`fa fa-caret-${docOpen ? 'down' : 'right'}`}
+											/>
+											<span className="gc-document-explorer-result-header-text">
+												{displayTitle}
+											</span>
+										</div>
+										<Collapse isOpened={docOpen}>
+											{doc.paragraphs.filter(paragraph => paragraph.paragraphIdBeingMatched === selectedInput).map((paragraph) =>{
+												let blockquoteClass = 'searchdemo-blockquote-sm';
+												const pOpen = selectedParagraph?.id === paragraph.id;
+												const isHighlighted = pOpen && docOpen;
+												if (isHighlighted)
+													blockquoteClass +=
+													' searchdemo-blockquote-sm-active';
+												return <div key={paragraph.id} style={{position: 'relative'}}>
+													{isHighlighted && (
+														<span className="searchdemo-arrow-left-sm"></span>
+													)}
+													<div
+														className={blockquoteClass}
+														onClick={(e) => {
+															e.preventDefault();
+															setCompareDocument(doc);
+															setSelectedParagraph(paragraph);
+														}}
+														style={{ 
+															marginLeft: 20, 
+															marginRight: 0,
+															border: isHighlighted ? 'none' : '1px solid #DCDCDC', 
+															padding: '3px',
+															cursor: 'pointer'
+														}}
+													>
+														<span className="gc-document-explorer-result-header-text" style={{color: isHighlighted ? 'white' : '#131E43' }}>
+															{isHighlighted ? `Page: ${paragraph.page_num_i + 1}, Par: ${paragraph.id.split('_')[1]}, Similarity Score: ${convertDCTScoreToText(paragraph.score)}` : paragraph.par_raw_text_t}
+														</span>
+													</div>
+													<Collapse isOpened={pOpen && docOpen}>
+														<div
+															className='searchdemo-blockquote-sm'
+															style={{ 
+																marginLeft: 20, 
+																marginRight: 0,
+																border: '1px solid #DCDCDC', 
+																padding: '10px',
+																whiteSpace: 'normal'
+															}}
+														>
+															<span className="gc-document-explorer-result-header-text" style={{fontWeight: 'normal'}}>
+																{paragraph.par_raw_text_t}
+															</span>
+															<div style={{display: 'flex', justifyContent:'right', marginTop:'10px'}}>
+																<GCTooltip title={'Export document matches to CSV'} placement="bottom" arrow>
+																	<GCButton
+																		onClick={() => exportSingleDoc(doc)}
+																		style={{marginLeft: 10, height: 36, padding: '0px, 10px', minWidth: 0, fontSize: '14px', lineHeight: '15px'}}
+																	>
+																		Export
+																	</GCButton>
+																</GCTooltip>
+																<GCTooltip title={'Save document to favorites'} placement="bottom" arrow>
+																	<GCButton 
+																		onClick={() => saveDocToFavorites(doc.filename, paragraph)}
+																		style={{marginLeft: 10, height: 36, padding: '0px, 10px', minWidth: 0, fontSize: '14px', lineHeight: '15px'}}
+																	>
+																		Save to Favorites
+																	</GCButton>
+																</GCTooltip>
+																<GCTooltip title={'Was this result relevant?'} placement="bottom" arrow>
+																	<i
+																		className={classes.feedback + ' fa fa-thumbs-up'}
+																		style={feedbackList[paragraph.id] ? {color: '#939395', WebkitTextStroke: '1px black'} : {}}
+																		onClick={() => handleFeedback(doc, paragraph, true)}
+																	/>
+																</GCTooltip>
+																<GCTooltip title={'Was this result relevant?'} placement="bottom" arrow>
+																	<i
+																		className={classes.feedback + ' fa fa-thumbs-down'}
+																		style={feedbackList[paragraph.id] === false ? {color: '#939395', WebkitTextStroke: '1px black'} : {}}
+																		onClick={() => handleFeedback(doc, paragraph, false)}
+																	/>
+																</GCTooltip>
+															</div>
+														</div>
+													</Collapse>
+												</div>
+											})}
+										</Collapse>
+									</div>
+								);
+							})}
+						</div>
+					</Grid>
+				</>
+				}
 			</Grid>
 		</>
 		
@@ -647,11 +960,9 @@ GCDocumentsComparisonTool.propTypes = {
 };
 
 const useStyles = makeStyles((theme) => ({
-	inputBoxRoot: {
-		backgroundColor: '#FFFFFF'
-	},
 	outlinedInput: {
 		color: '#0000008A',
+		backgroundColor: '#FFFFFF',
 		fontFamily: 'Montserrat',
 		fontSize: 14,
 		height: 247,
@@ -662,6 +973,30 @@ const useStyles = makeStyles((theme) => ({
 		},
 		'& textarea': {
 			height: '100%'
+		}
+	},
+	root: {
+		paddingTop: '16px',
+		marginRight: '10px',
+		'& .MuiInputBase-root':{
+			height: '50px',
+			fontSize: 20
+		},
+		'& .MuiFormLabel-root': {
+			fontSize: 20
+		},
+		'&:hover .MuiInput-underline:before':{
+			borderBottom: `3px solid ${gcOrange}`
+		},
+		'& .MuiInput-underline:before':{
+			borderBottom: `3px solid rgba(0, 0, 0, 0.42)`
+		},
+		'& .MuiInput-underline:after':{
+			borderBottom: `3px solid ${gcOrange}`
+		},
+		'& .Mui-focused':{
+			borderColor: `${gcOrange}`,
+			color:`${gcOrange}`
 		}
 	},
 	focused: {},
@@ -676,9 +1011,33 @@ const useStyles = makeStyles((theme) => ({
 		backgroundColor: '#EFF1F6',
 		height: 850
 	},
+	filterBox: {
+		backgroundColor: '#ffffff',
+		borderRadius: '5px',
+		padding: '2px',
+		border: '2px solid #bdccde',
+		pointerEvents: 'none',
+		margin: 'auto 5px auto 0px',
+		height: 19,
+		width: 19
+	},
+	feedback: {
+		fontSize: 25,
+		margin: 'auto 0 auto 10px',
+		justifyContent: 'center',
+		color: 'white',
+		'-webkit-text-stroke': '1px #808080',
+		'&:hover': {
+			cursor: 'pointer',
+			'-webkit-text-stroke': '1px black',
+		},
+	},
+	selectIcon: {
+		marginTop: '4px'
+	},
+	formlabel: {
+		paddingTop: '16px'
+	}
 }));
 
 export default GCDocumentsComparisonTool;
-/*	
-style={{ border: 'none', backgroundColor: gcOrange, padding: '0 15px', display: 'flex', height: 50, alignItems: 'center', borderRadius: 5, margin: '0 0 0 1150px'}}				
-*/
